+++ packages/eslint-config-nikita/index.js
@@ -1,12 +1,16 @@
 module.exports = {
-    extends: ['./rules/best-practices', './rules/errors', './rules/es6', './rules/strict', './rules/style', './rules/variables', './rules/imports'].map(require.resolve),
+    extends: [
+        './rules/best-practices',
+        './rules/errors',
+        './rules/es6',
+        './rules/strict',
+        './rules/style',
+        './rules/variables',
+        './rules/imports',
+    ].map(require.resolve),
     parserOptions: {
-        ecmaVersion: 2017,
+        ecmaVersion: 2018,
         sourceType: 'module',
-        ecmaFeatures: {
-            experimentalObjectRestSpread: true,
-        },
     },
-    rules: {
-    },
+    rules: {},
 };